added webpack and eslint , now it works with webpack, and it's not longer just 3 files linked
