--- index.html
@@ -1,124 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>My ToDo Lists</title>
-    <link rel="stylesheet" type="text/css" href="styles.css">
-    <script defer src="main.js"></script>
-
-</head>
-<body>
-    <h1 class="title">Things to do</h1>
-
-    <div class="all-tasks">
-
-        <h2 class="task-list-title">My lists</h2>
-        <ul class="task-list" data-lists="">
-            <!-- <li  class='list-name 
-            active-list'>Todo code</li>
-            <li class='list-name'>Work</li>
-            <li class='list-name'>Groceries</li> -->
-        </ul>
-
-        <form action="" data-new-list-form >
-            <input 
-            type="text" 
-            class="new list"
-            data-new-list-input
-            placeholder="new list name"
-            aria-label="new list name" 
-            >
-            <button class="btn create"
-            aria-label="create new list">+</button>
-        </form>
-
-    </div>
-
-    <div class="todo-list" data-list-display-container>
-        <div class="todo-header">
-            <h2 class="list-title" data-list-title >Todo Code</h2>
-            <p class="task-count" data-list-count > 3 task remaining</p>
-        </div>
-
-
-        <div class="todo-body">
-
-            <div class="tasks" data-tasks>
-
-                <!-- task 1
-                <div class="task">
-                    <input 
-                    type="checkbox" 
-                    id="task-1">
-                    <label for="task-1">
-                        <span class="custom-check-box"></span>
-                        this is a task
-                    </label>
-                </div>
-
-                task 2 
-
-                <div class="task">
-                    <input 
-                    type="checkbox" 
-                    id="task-2">
-                    <label for="task-2">
-                        <span class="custom-check-box"></span>
-                        this is another task
-                    </label>
-                </div>
-
-               
-                <div class="task">
-                    <input 
-                    type="checkbox" 
-                    id="task-3">
-                    <label for="task-3">
-                        <span class="custom-check-box"></span>
-                        this is a third task
-                    </label>
-                </div> -->
-
-
-            </div>
-
-            <div class="new-task-creator" data-new-task-form >
-
-                <form action="">
-                    <input 
-                    type="text" 
-                    class="new task"
-                    data-new-task-input
-                    placeholder="new task name"
-                    aria-label="new task name" 
-                    >
-                    <button class="btn create"
-                    aria-label="create new task">+</button>
-                </form>
-
-            </div>
-            
-            <div class="delete-stuff">
-                <button class="btn delete" data-clear-complete-task-buttton > Clear completed task</button>
-                <button class="btn delete" data-delete-list--button>Delete list</button>
-            </div>
-
-
-        </div>
-
-
-    </div>
-
-    <template id='task-template'>
-        <div class="task">
-            <input 
-            type="checkbox">
-            <label for="task-1">
-                <span class="custom-check-box"></span>
-            </label>
-        </div>
-    </template>
-
-</body>
-</html>
